refactor(taxa-autocomplete): split into view + container
Extract pure presentation into TaxaAutocompleteView (options/loading/onSearch
as props) and reduce TaxaAutocomplete to a thin wrapper around useAutocomplete.
Story variants (results, loading, empty) now live on the view and stub data
directly without MSW or play-function workarounds.
